% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpklips.R
\name{getpklips}
\alias{getpklips}
\title{Obtain data.frame for the Korea Labor & Income Panel Study (only household member survey)}
\usage{
getpklips(
  path,
  year,
  datatype = c("stata", "spss", "sas", "xlsx"),
  klipsvars = c("0101", "0107"),
  outvars = c("gender", "age")
)
}
\arguments{
\item{path}{A string specifying folder containing KLIPS household member survey data}

\item{year}{integer specifying the years from 1998 to 2023 that the user wants to include in the dataframe.}

\item{datatype}{A string specifying the format of the raw data you want to convert to a data frame ("spss", "sas", "stata", "excel")}

\item{klipsvars}{A string or strings specifying the variables in the raw data that you want to convert to a data frame ("0101", "0107")}

\item{outvars}{A string or strings specifying the variable names of converted data ("gender", "age")}
}
\value{
A data frame containing klips household member data with the specified years and variables.
}
\description{
Obtain data.frame for the Korea Labor & Income Panel Study (only household member survey)
}
\examples{
\dontrun{
# Plot of race/ethnicity by county in Illinois for 2010
library(tidyklips)
library(tidyverse)

df <- getpklips(path = "../datastata", year = 2022:2023, datatype = "stata")
df \%>\%
  group_by(year, gender) \%>\%
  summarise(count = n(),
            proportion = count/sum(gender, na.rm=TRUE) )
}
}
