% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klips26p.R
\docType{data}
\name{klips26p}
\alias{klips26p}
\title{2023 yr Klips house member survey dataframe}
\format{
Data frame with columns
\describe{
\item{pid}{response id.}
\item{gender}{gender.}
\item{age}{age.}
}
}
\source{
\url{https://www.kli.re.kr/klips}
}
\usage{
klips26p
}
\description{
Look up 23364 house member survey response data.
}
\examples{
klips26p
}
\keyword{datasets}
