% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyklips-package.R
\docType{package}
\name{tidyklips-package}
\alias{tidyklips}
\alias{tidyklips-package}
\title{tidyklips: Load Korea Labor & Income Panel Study (KLIPS) Data as Data Frames}
\description{
Loading the Korea Labor Institute's KLIPS (Korea Labor & Income Panel Study) panel data and returning data frames. Users must download 26 years of panel data from the Korea Labor Institute website and save it in a folder in an appropriate path. The data can be downloaded in stata, spss, sas, or xlsx formats. Afterwards, users can easily convert the data into a data frame using this package.
}
\author{
\strong{Maintainer}: Cheon Geun Choi \email{cheongeun.choi@gmail.com}

}
\keyword{internal}
